+++ lib/api/SetAttribute.ts
@@ -15,7 +15,12 @@
  */
 import { ContextModel } from '../models/user/ContextModel';
 import { EventEnum } from '../enums/EventEnum';
-import { getEventsBaseProperties, getAttributePayloadData, sendPostApiRequest } from '../utils/NetworkUtil';
+import {
+  getEventsBaseProperties,
+  getAttributePayloadData,
+  sendPostApiRequest,
+  getShouldWaitForTrackingCalls,
+} from '../utils/NetworkUtil';
 import { SettingsModel } from '../models/settings/SettingsModel';
 
 interface ISetAttribute {
@@ -26,7 +31,12 @@ interface ISetAttribute {
    * @param attributeValue The value of the attribute.
    * @param context Context containing user information.
    */
-  setAttribute(settings: SettingsModel, attributeKey: string, attributeValue: any, context: ContextModel): void;
+  setAttribute(
+    settings: SettingsModel,
+    attributeKey: string,
+    attributeValue: any,
+    context: ContextModel,
+  ): Promise<void>;
 }
 
 export class SetAttributeApi implements ISetAttribute {
@@ -37,8 +47,17 @@ export class SetAttributeApi implements ISetAttribute {
    * @param attributeValue The value of the attribute.
    * @param context Context containing user information.
    */
-  setAttribute(settings: SettingsModel, attributeKey: string, attributeValue: any, context: ContextModel): void {
-    createImpressionForAttribute(settings, attributeKey, attributeValue, context);
+  async setAttribute(
+    settings: SettingsModel,
+    attributeKey: string,
+    attributeValue: any,
+    context: ContextModel,
+  ): Promise<void> {
+    if (getShouldWaitForTrackingCalls()) {
+      await createImpressionForAttribute(settings, attributeKey, attributeValue, context);
+    } else {
+      createImpressionForAttribute(settings, attributeKey, attributeValue, context);
+    }
   }
 }
 
@@ -74,5 +93,5 @@ const createImpressionForAttribute = async (
   );
 
   // Send the constructed payload via POST request
-  sendPostApiRequest(properties, payload);
+  await sendPostApiRequest(properties, payload);
 };