+++ lib/api/TrackEvent.ts
@@ -22,7 +22,12 @@ import IHooksService from '../services/HooksService';
 import { dynamic } from '../types/Common';
 import { doesEventBelongToAnyFeature } from '../utils/FunctionUtil';
 import { buildMessage } from '../utils/LogMessageUtil';
-import { getEventsBaseProperties, getTrackGoalPayloadData, sendPostApiRequest } from '../utils/NetworkUtil';
+import {
+  getEventsBaseProperties,
+  getTrackGoalPayloadData,
+  sendPostApiRequest,
+  getShouldWaitForTrackingCalls,
+} from '../utils/NetworkUtil';
 
 interface ITrack {
   /**
@@ -57,7 +62,11 @@ export class TrackApi implements ITrack {
   ): Promise<Record<string, boolean>> {
     if (doesEventBelongToAnyFeature(eventName, settings)) {
       // Create an impression for the track event
-      createImpressionForTrack(settings, eventName, context, eventProperties);
+      if (getShouldWaitForTrackingCalls()) {
+        await createImpressionForTrack(settings, eventName, context, eventProperties);
+      } else {
+        createImpressionForTrack(settings, eventName, context, eventProperties);
+      }
       // Set and execute integration callback for the track event
       hooksService.set({ eventName: eventName, api: ApiEnum.TRACK });
       hooksService.execute(hooksService.get());
@@ -105,5 +114,5 @@ const createImpressionForTrack = async (
     context?.getIpAddress(),
   );
   // Send the prepared payload via POST API request
-  sendPostApiRequest(properties, payload);
+  await sendPostApiRequest(properties, payload);
 };