+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.1] - 2024-09-03
+
+### Added
+
+- feat: added support to wait for network response incase of edge like environment.
+
+  ```javascript
+  const { init } = require('vwo-fme-node-sdk');
+
+  const vwoClient = await init({
+    accountId: '123456', // VWO Account ID
+    sdkKey: '32-alpha-numeric-sdk-key', // SDK Key
+    shouldWaitForTrackingCalls: true, // if running on edge env
+  });
+  ```
+
 ## [1.6.0] - 2024-08-27
 
 ### Fixed